+++ test/test_mortoncurve4.cc
@@ -169,7 +169,7 @@ int main(){
 		using namespace morton_curve;
 
 		for(auto const &[zzz, x, y, z, w] : container){
-			printf("%6lu %4lu %4lu %4lu %4lu\n", zzz, x, y, z, w);
+			printf("%6lu %4u %4u %4u %4u\n", (uint64_t)zzz, x, y, z, w);
 		}
 	}
